move deviceListItem and recoverMethods to a component
